Fix for MSVC warnings.

+++ xcode_unique_id.cpp
@@ -24,12 +24,12 @@
 #include <ctime>
 #include <sstream>
 
-static volatile size_t g_Counter = 1;
+static volatile unsigned int g_Counter = 1;
 
 XCodeUniqueID::XCodeUniqueID()
 {
-	m_ID[0] = time(NULL);
-	m_ID[1] = reinterpret_cast<size_t>(this) & 0xFFFFFFFFU;
+	m_ID[0] = static_cast<unsigned int>(time(NULL));
+	m_ID[1] = static_cast<unsigned int>(reinterpret_cast<size_t>(this) & 0xFFFFFFFFU);
 	m_ID[2] = ((m_ID[0] + m_ID[1]) ^ 0xCAFEBABE) + g_Counter++;
 }
